+++ CHANGES.md
@@ -10,6 +10,7 @@ v1.1.0 - YYYY-MM-DD
 - Added `ttfCreateData` function to create a font from a memory buffer.
 - Added `ttfContainsChar` and `ttfContainsChars` functions to determine whether
   a font contains the necessary characters (Issue #2)
+- Added `ttfGetKernedExtents` function to get kerned extents (Issue #6)
 - Added guards against fonts claiming they have 0 characters.
 - Added stddef.h include to ttf.h.
 - Added support for more kinds of TrueType/OpenType fonts.